+++ backend/src/services/game.service.ts
@@ -1,6 +1,57 @@
 import prisma from "../db";
-import {Game, Player, Season} from "@prisma/client";
-import {CreateGameType} from "../validation/game.validation";
+import {Game, GameType, GameVariant, Player} from "@prisma/client";
+
+const createGame = async (gameVariant: GameVariant, gameType: GameType, players: any[], defaultRound: any, recorderId: string, seasonId: string): Promise<Game> => {
+    return prisma.game.create({
+        data: {
+            season: {
+                connect: {
+                    id: seasonId
+                }
+            },
+            gameVariant: gameVariant,
+            gameType: gameType,
+            status: 'IN_PROGRESS',
+            recordedBy: {
+                connect: {
+                    id: recorderId
+                }
+            },
+            players: {
+                create: players
+            },
+            japaneseRounds: {
+                create: [{
+                    roundCount: 1,
+                    roundNumber: 1,
+                    roundWind: 'EAST',
+                    bonus: 0,
+                    riichiSticks: 0
+                }]
+            },
+            [gameVariant === GameVariant.JAPANESE ? 'japaneseRounds' : 'hongKongRounds']: {
+                create: defaultRound
+            }
+        }
+    });
+}
+
+const getGame = async (gameId: number): Promise<any> => {
+    return prisma.game.findUnique({
+        where: {
+            id: gameId
+        },
+        include: {
+            players: {
+                include: {
+                    player: true
+                }
+            },
+            japaneseRounds: true,
+            hongKongRounds: true
+        }
+    })
+}
 
 // Throws error if the player list contains duplicates
 const checkPlayerListUnique = (playerList: string[]): void => {
@@ -20,37 +71,38 @@ const checkPlayerGameEligibility = (gameVariant: string, player: Player): void =
     throw new Error("Player not eligible for game type")
 }
 
-const createGame = async (createGame: CreateGameType, players: any[], recorderId: string, seasonId: string): Promise<Game> => {
-    return prisma.game.create({
-        data: {
-            season: {
-                connect: {
-                    id: seasonId
-                }
-            },
-            gameType: createGame.gameType,
-            status: 'IN_PROGRESS',
-            recordedBy: {
-                connect: {
-                    id: recorderId
-                }
-            },
-            players: {
-                create: players
-            },
-            japaneseGames: {
-                create: [{
-                    rounds: {
-                        create: {
-                            roundCount: 1,
-                            roundNumber: 1,
-                            roundWind: 'EAST'
-                        }
-                    }
-                }]
-            }
-        }
-    });
+const getWind = (index: number): string => {
+    switch (index) {
+        case 0:
+            return "EAST"
+        case 1:
+            return "SOUTH"
+        case 2:
+            return "WEST"
+        case 3:
+            return "NORTH"
+        default:
+            return "NONE"
+    }
 }
 
-export {checkPlayerGameEligibility, createGame, checkPlayerListUnique}
+const getDefaultRound = (gameVariant: GameVariant): any => {
+    if (gameVariant === GameVariant.JAPANESE) {
+        return {
+            roundCount: 1,
+            roundNumber: 1,
+            roundWind: 'EAST',
+            bonus: 0,
+            riichiSticks: 0
+        }
+    } else if (gameVariant === GameVariant.HONG_KONG) {
+        return {
+            roundCount: 1,
+            roundNumber: 1,
+            roundWind: 'EAST',
+            bonus: 0
+        }
+    }
+}
+
+export {checkPlayerGameEligibility, checkPlayerListUnique, getWind, getDefaultRound, createGame, getGame}