+++ frontend/src/game/CreateGame.tsx
@@ -5,8 +5,11 @@ import {AxiosError} from "axios";
 import {AuthContext} from "../common/AuthContext";
 import {withPlayerCondition} from "../common/withPlayerCondition";
 import Select from 'react-select'
+import {useNavigate} from "react-router-dom";
+import {getGameTypeString, getGameVariant} from "../common/Utils";
 
 const CreateGameComponent: FC<GameTypeProp> = ({gameVariant}) => {
+    const navigate = useNavigate();
     const { player } = useContext(AuthContext);
 
     const [playerNames, setPlayerNames] = useState<string[]>([])
@@ -36,28 +39,10 @@ const CreateGameComponent: FC<GameTypeProp> = ({gameVariant}) => {
 
         const playerList = [eastPlayer, southPlayer, westPlayer, northPlayer];
         createGameAPI(player!.authToken, "RANKED", getGameVariant(gameVariant), playerList).then((response) => {
-            console.log("Created game: ", response.data)
+            navigate(`/games/${response.data.id}`)
         }).catch((error: AxiosError) => {
             console.log("Error creating game: ", error.response?.data)
         })
-    }
-
-    const getGameTypeString = (gameType: "jp" | "hk"): string => {
-        if (gameType === "jp") {
-            return "Riichi"
-        } else if (gameType === "hk") {
-            return "Hong Kong"
-        }
-        return ""
-    }
-
-    const getGameVariant = (gameType: "jp" | "hk"): GameVariant => {
-        if (gameType === "jp") {
-            return "JAPANESE"
-        } else if (gameType === "hk") {
-            return "HONG_KONG"
-        }
-        return "JAPANESE"
     }
 
     const title = `Create Ranked ${getGameTypeString(gameVariant)} Game`