+++ frontend/src/type.d.ts
@@ -58,3 +58,18 @@ type AuthContextType = {
     register: (registerData: RegisterDataType) => Promise<void>;
     logout: () => Promise<void>;
 };
+
+type Game = {
+    id: string
+    gameType: GameType
+    gameVariant: GameVariant
+    status: string
+    recordedById: string
+    players: any[]
+    rounds: any[]
+}
+
+type LeaderboardType = {
+    username: string
+    elo: number
+}