+++ backend/CHANGELOG.md
@@ -7,5 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2023-05-04
+
+### Added
+- Leaderboard routes for backend
+
+## [0.0.2] - 2023-05-02
+
+### Added
+- getGame routes for backend
+
+### Changed
+- Simplified Game schema
+
 ## [0.0.1] - 2023-04-17
 - Start a changelog